+++ src/wrappers/themis/react-native-themis/src/index.js
@@ -226,7 +226,7 @@ export function secureCellContextImprintDecrypt64(symmetricKey64, encrypted64, c
     });
 }
 // secure message sign and verify
-export function secureMessageSign64(plaintext, privateKey64, publicKey64 = "") {
+export function secureMessageSign64(plaintext, privateKey64) {
     if (plaintext === "" || plaintext === undefined || plaintext === null) {
         throw new Error("Parameter plaintext can not be empty");
     }
@@ -237,15 +237,8 @@ export function secureMessageSign64(plaintext, privateKey64, publicKey64 = "") {
         throw new Error("Parameter privateKey64 is not base64 encoded");
     }
     const privateKey = Array.from(Buffer.from(privateKey64, 'base64'));
-    if (publicKey64 === undefined || publicKey64 === null) {
-        publicKey64 = "";
-    }
-    if (publicKey64 && !isBase64(publicKey64)) {
-        throw new Error("Optional parameter publicKey64 is not base64 encoded");
-    }
-    const publicKey = publicKey64 === "" ? null : Array.from(Buffer.from(publicKey64, 'base64'));
     return new Promise((resolve, reject) => {
-        Themis.secureMessageSign(plaintext, privateKey, publicKey, (signed) => {
+        Themis.secureMessageSign(plaintext, privateKey, (signed) => {
             resolve(Buffer.from(new Uint8Array(signed)).toString("base64"));
         }, (error) => {
             reject(error);